use new rpc stream method to watch vpn status
