+++ nym-vpn-x/src-tauri/src/grpc/client.rs
@@ -6,7 +6,7 @@ use nym_vpn_proto::{
     nym_vpnd_client::NymVpndClient, DisconnectRequest, HealthCheckRequest, StatusRequest,
 };
 use nym_vpn_proto::{
-    ConnectRequest, ConnectionStatus, Dns, EntryNode, ExitNode, ImportUserCredentialRequest,
+    ConnectRequest, ConnectionStatus, Dns, Empty, EntryNode, ExitNode, ImportUserCredentialRequest,
 };
 use parity_tokio_ipc::Endpoint as IpcEndpoint;
 use serde::{Deserialize, Serialize};
@@ -18,6 +18,7 @@ use tonic::{transport::Channel, Request};
 use tracing::{debug, error, info, instrument, warn};
 use ts_rs::TS;
 
+use super::vpn_status::vpn_status_update;
 use crate::cli::Cli;
 use crate::fs::config::AppConfig;
 use crate::{events::AppHandleEventEmitter, states::SharedAppState};
@@ -129,6 +130,7 @@ impl GrpcClient {
         Ok(status.into())
     }
 
+    /// Get VPN status
     #[instrument(skip_all)]
     pub async fn vpn_status(&self) -> Result<ConnectionStatus, VpndError> {
         debug!("vpn_status");
@@ -142,6 +144,46 @@ impl GrpcClient {
         debug!("grpc response: {:?}", response);
 
         Ok(response.into_inner().status())
+    }
+
+    /// Watch VPN status updates
+    #[instrument(skip_all)]
+    pub async fn watch_vpn_status(&self, app: &AppHandle) -> Result<()> {
+        let mut vpnd = self.vpnd().await?;
+
+        let request = Request::new(Empty {});
+        let mut stream = vpnd
+            .listen_to_connection_state_changes(request)
+            .await
+            .inspect_err(|e| {
+                error!("listen_to_connection_state_changes failed: {}", e);
+            })?
+            .into_inner();
+
+        let (tx, mut rx) = mpsc::channel(32);
+        tokio::spawn(async move {
+            loop {
+                match stream.message().await {
+                    Ok(Some(update)) => {
+                        tx.send(update).await.unwrap();
+                    }
+                    Ok(None) => {
+                        warn!("watch vpn status stream closed by the server");
+                        return;
+                    }
+                    Err(e) => {
+                        warn!("watch vpn status stream get a grpc error: {}", e);
+                    }
+                }
+            }
+        });
+
+        while let Some(status) = rx.recv().await {
+            debug!("vpn status update {:?}", status);
+            vpn_status_update(app, status).await?;
+        }
+
+        Ok(())
     }
 
     /// Connect to the VPN
