+++ nym-vpn-x/src-tauri/src/grpc/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+mod vpn_status;
